Added validation. Also clean-up.

+++ trad4/bin/recreate_db.sh
@@ -1,31 +1,41 @@
 #!/bin/sh
 
+SQL=$TRAD4_ROOT/sql
+GSQL=$APP_ROOT/gen/sql
+LSQL=$TRAD4_ROOT/bin/lsql
+
+if [ ! -d $GSQL ]
+then
+    echo "Error: $GSQL not found. Have you run the precompiler?"
+    exit 1
+fi
+
 if [ -f $APP_DB ]
 then
     echo "Dropping $APP_DB..."
     rm $APP_DB
 fi
 
-echo "Loading all in $TRAD4_ROOT/sql..."
+echo "Loading all in $SQL..."
 
-cd $TRAD4_ROOT/sql
+cd $SQL
 
 for i in `ls *.table`
 do
     echo "Loading $i"
 
-    $TRAD4_ROOT/bin/lsql $i
+    $LSQL $i
 done
 
-echo "Loading all in $APP_ROOT/gen/sql..."
+echo "Loading all in $GSQL..."
 
-cd $APP_ROOT/gen/sql
+cd $GSQL
 
 for i in `ls *.table`
 do
     echo "Loading $i"
 
-    $TRAD4_ROOT/bin/lsql $i
+    $LSQL $i
 done
